Moved color-related components to `ColorPreferences.ts`
